Testing (#6)
* Add gems for testing

- Added the Thor gem to make running tests easier; thor was also needed
for adding DB migrations, which were required for the first test to run
- Added test-unit gem since Bundler requires it as a gem to be found
properly by `require`
- test-unit does not come with mocking or stubbing so Mocha was added
to accomodate those needs

* Explicitely ignore sqlite3 ext rather than db

I've chosen to use a specific sqlite3 extension for databases, akin to
Rails.

* Add DB directory

* Structure config like Rails

As an experiment, I've structured the config directory load hierarchy
like Rails. Time will tell if this is a smart decision or not.

* Loader deleted

It was garbage and untestible. Removed.

* Base testing infrastructure

- the `test_helper` is meant to be included in every test
- the `all` file is for test/unit to run all tests as a suite

* Test that the user model exists

* Add a thor CLI command

- This probably needs refinement but it works
- It should get commands for launching the server
- It runs migrations

* Update rest of app with new structure

The Rails convention of storing eventhing to bootstrap the app in
the config/ path makes for a much cleaner elsewhere in the app.
The root application, rack config and models helper were all
cleaned up.