Test application (#7)
* Include files in the proper locations

Schema depends on types and the Graphql endpoint depends on the
Schema. Require files in the appropirate dependency graph.

* Abstract template helper

It seems insane to include the entire application just to process an ERB
template. It makes far more sense to create a helper to contain the
code needed by the template and bind it.

* Test User a little more thoroughly

* Restructure Graphql endpoint for testing

* Change load path for tests to root
- This forces more explicit requires, for example `modules/user`.

* Test graphql route

* Add test for graphiql endpoint

* Fix env for routing in test and resulting bug

The change to the graphql route actually introduced a bug. The fixed
test means a fixed route

* Add tests for posting to endpoints